+++ metrics/crowdsec/active-bans.test.ts
@@ -1,0 +1,5 @@
+import { expectFixtureMetric } from '../test-utils'
+
+it('counts active CrowdSec bans', async () => {
+  await expectFixtureMetric(new URL('./active-bans.yml', import.meta.url), [{}, {}], 2)
+})